Create data details.txt
In 1912, the ship RMS Titanic struck an iceberg on its maiden voyage and sank, resulting in the deaths of most of its passengers and crew. In this project, it will explore a subset of the RMS Titanic passenger manifest to determine which features best predict whether someone survived or did not survive. To achieve accurate prediction, several conditional predictions are implemented using different machine learning models.